refactor: author component
